podcastify_service/requirements.txt
podcastfy>=0.4.0
ffmpeg-python>=0.2.0
pandas>=2.2.3
python-dotenv==1.0.1
requests>=2.31.0
flask==3.1.0 